+++ nw/MAIN/RecoveredSoftwareFrame.h
@@ -1,0 +1,6 @@
+#ifndef RR2NW_RECOVERED_SOFTWARE_FRAME_H
+#define RR2NW_RECOVERED_SOFTWARE_FRAME_H
+
+void Frame_BindRecoveredSoftware();
+
+#endif  // RR2NW_RECOVERED_SOFTWARE_FRAME_H