+++ nw/SCENE/VesselRuntimeGlobals.cpp
@@ -1,0 +1,3 @@
+class CViewObjectRef;
+
+CViewObjectRef* pVesselObj = 0;