+++ nw/VIEW/TerrainFrameCounters.cpp
@@ -1,0 +1,2 @@
+int __revCnt = 0;
+int __fullCnt = 0;